Add the LocalDateCompare class
