+++ api-java/src/main/java/javaclasses/mealorder/c/aggregate/VendorAggregate.java
@@ -18,13 +18,12 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
 
-package javaclasses.mealorder.c.aggregate.aggregate;
+package javaclasses.mealorder.c.aggregate;
 
 import com.google.protobuf.Message;
 import io.spine.server.aggregate.Aggregate;
 import io.spine.server.aggregate.Apply;
 import io.spine.server.command.Assign;
-import io.spine.time.LocalDate;
 import javaclasses.mealorder.MenuDateRange;
 import javaclasses.mealorder.Vendor;
 import javaclasses.mealorder.VendorId;
@@ -38,14 +37,13 @@ import javaclasses.mealorder.c.event.MenuImported;
 import javaclasses.mealorder.c.event.VendorAdded;
 import javaclasses.mealorder.c.event.VendorUpdated;
 import javaclasses.mealorder.c.rejection.CannotSetDateRange;
-import javaclasses.mealorder.c.rejection.VendorAlreadyExists;
 
 import java.util.List;
 
 import static io.spine.time.Time.getCurrentTime;
 import static java.util.Collections.singletonList;
-import static javaclasses.mealorder.c.aggregate.aggregate.VendorFlowValidator.isValidDateRange;
-import static javaclasses.mealorder.c.aggregate.aggregate.rejection.VendorAggregateRejections.UpdateRejections.throwCannotSetDateRange;
+import static javaclasses.mealorder.c.aggregate.VendorValidator.isValidDateRange;
+import static javaclasses.mealorder.c.aggregate.rejection.VendorAggregateRejections.UpdateRejections.throwCannotSetDateRange;
 
 /**
  * The aggregate managing the state of a {@link Vendor}.
