+++ api-java/src/main/java/javaclasses/mealorder/c/repository/package-info.java
@@ -1,7 +1,7 @@
 //
 // Copyright 2018, TeamDev Ltd. All rights reserved.
 //
-// Redistribution and use in source and/or binary gorms, with or without
+// Redistribution and use in source and/or binary forms, with or without
 // modification, must retain the above copyright notice and the following
 // disclaimer.
 //
@@ -19,10 +19,9 @@
 //
 
 /**
- * This package contains implementation for classes which provides
- * methods to process aggregates and projection.
+ * This package provides classes for working with aggregate and projection repositories.
  */
 @ParametersAreNonnullByDefault
-package javaclasses.mealorder.c.aggregate;
+package javaclasses.mealorder.c.repository;
 
 import javax.annotation.ParametersAreNonnullByDefault;
