+++ api-java/src/test/java/javaclasses/mealorder/c/definition/AddVendorTest.java
@@ -18,7 +18,7 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
 
-package javaclasses.mealorder.c.aggregate.definition;
+package javaclasses.mealorder.c.definition;
 
 import com.google.protobuf.Message;
 import javaclasses.mealorder.Vendor;
