+++ api-java/src/test/java/javaclasses/mealorder/c/definition/UpdateVendorTest.java
@@ -18,14 +18,13 @@
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
 
-package javaclasses.mealorder.c.aggregate.definition;
+package javaclasses.mealorder.c.definition;
 
 import com.google.protobuf.Message;
 import javaclasses.mealorder.Vendor;
 import javaclasses.mealorder.c.command.AddVendor;
-import javaclasses.mealorder.c.command.ImportMenu;
-import javaclasses.mealorder.c.command.SetDateRangeForMenu;
-import javaclasses.mealorder.c.event.DateRangeForMenuSet;
+import javaclasses.mealorder.c.command.UpdateVendor;
+import javaclasses.mealorder.c.event.VendorUpdated;
 import javaclasses.mealorder.testdata.TestVendorCommandFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -34,8 +33,7 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 
 import static io.spine.server.aggregate.AggregateMessageDispatcher.dispatchCommand;
-import static javaclasses.mealorder.testdata.TestVendorCommandFactory.MENU_DATE_RANGE;
-import static javaclasses.mealorder.testdata.TestVendorCommandFactory.MENU_ID;
+import static javaclasses.mealorder.testdata.TestVendorCommandFactory.NEW_VENDOR_NAME;
 import static javaclasses.mealorder.testdata.TestVendorCommandFactory.USER_ID;
 import static javaclasses.mealorder.testdata.TestVendorCommandFactory.VENDOR_ID;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -44,8 +42,8 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 /**
  * @author Yurii Haidamaka
  */
-@DisplayName("ImportMenu command should be interpreted by VendorAggregate and")
-public class SetDateRangeForMenuTest extends VendorCommandTest<AddVendor> {
+@DisplayName("UpdateVendor command should be interpreted by VendorAggregate and")
+public class UpdateVendorTest extends VendorCommandTest<AddVendor> {
 
     @Override
     @BeforeEach
@@ -54,33 +52,32 @@ public class SetDateRangeForMenuTest extends VendorCommandTest<AddVendor> {
     }
 
     @Test
-    @DisplayName("produce DateRangeForMenuSet event")
+    @DisplayName("produce UpdateVendor event")
     void produceEvent() {
-        final SetDateRangeForMenu setDateRangeForMenuCmd = TestVendorCommandFactory.setDateRangeForMenuInstance();
+        final UpdateVendor updateVendorCmd = TestVendorCommandFactory.updateVendorInstance();
 
         final List<? extends Message> messageList = dispatchCommand(aggregate,
-                                                                    envelopeOf(
-                                                                            setDateRangeForMenuCmd));
+                                                                    envelopeOf(updateVendorCmd));
 
         assertNotNull(aggregate.getId());
         assertEquals(1, messageList.size());
-        assertEquals(DateRangeForMenuSet.class, messageList.get(0)
-                                                           .getClass());
-        final DateRangeForMenuSet dateRangeForMenuSet = (DateRangeForMenuSet) messageList.get(0);
+        assertEquals(VendorUpdated.class, messageList.get(0)
+                                                     .getClass());
+        final VendorUpdated vendorUpdated = (VendorUpdated) messageList.get(0);
 
-        assertEquals(VENDOR_ID, dateRangeForMenuSet.getVendorId());
-        assertEquals(MENU_ID, dateRangeForMenuSet.getMenuId());
-        assertEquals(USER_ID, dateRangeForMenuSet.getWhoSet());
-        assertEquals(MENU_DATE_RANGE, dateRangeForMenuSet.getMenuDateRange());
+        assertEquals(VENDOR_ID, vendorUpdated.getVendorId());
+        assertEquals(USER_ID, vendorUpdated.getWhoUploaded());
+        assertEquals(NEW_VENDOR_NAME, vendorUpdated.getVendorChange()
+                                                   .getNewVendorName());
     }
 
     @Test
-    @DisplayName("set date range for menu")
+    @DisplayName("update vendor")
     void addVendor() {
-        final ImportMenu importMenu = TestVendorCommandFactory.importMenuInstance();
-        dispatchCommand(aggregate, envelopeOf(importMenu));
+        final UpdateVendor updateVendor = TestVendorCommandFactory.updateVendorInstance();
+        dispatchCommand(aggregate, envelopeOf(updateVendor));
 
         final Vendor state = aggregate.getState();
-        assertEquals(state.getId(), importMenu.getVendorId());
+        assertEquals(state.getId(), updateVendor.getVendorId());
     }
 }
